Tidy up package structure
